implements changes from a9878aee1b6f3ac5df67f22b2a8de8248093cc67 to 99ceef0033aa56cd5d89db751732a5ceb18fc434

+++ components/actions/FileDelete.js
@@ -10,7 +10,8 @@ export default async function deleteFile(execute = false, act, template) {
                 return { warning: `Target path does not exist.`, data };
         if (!execute)
             return { data };
-        fs.unlinkSync(data.target);
+        if (fs.existsSync(data.target))
+            fs.unlinkSync(data.target);
         return { success: true, data };
     }
     catch (e) {
